Refactor and add search link to navbar
